CSS Variables refactor - untitledTextEditor
