+++ src/vs/workbench/contrib/codeEditor/browser/untitledTextEditorHint/untitledTextEditorHint.css
@@ -1,0 +1,12 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+.monaco-editor .contentWidgets .untitled-hint {
+	color: var(--vscode-input-placeholderForeground);
+}
+
+.monaco-editor .contentWidgets .untitled-hint a {
+	color: var(--vscode-textLink-foreground)
+}
